aws_ed_api_client/models/enrollment_json.py
from typing import Any, Dict, List, Type, TypeVar, Union

import attr

from ..types import UNSET, Unset

T = TypeVar("T", bound="EnrollmentJson")


@attr.s(auto_attribs=True)
class EnrollmentJson:
    """
    Attributes:
        username (Union[Unset, str]):
        course (Union[Unset, str]):
    """

    username: Union[Unset, str] = UNSET
    course: Union[Unset, str] = UNSET
    additional_properties: Dict[str, Any] = attr.ib(init=False, factory=dict)

    def to_dict(self) -> Dict[str, Any]:
        username = self.username
        course = self.course

        field_dict: Dict[str, Any] = {}
        field_dict.update(self.additional_properties)
        field_dict.update({})
        if username is not UNSET:
            field_dict["username"] = username
        if course is not UNSET:
            field_dict["course"] = course

        return field_dict

    @classmethod
    def from_dict(cls: Type[T], src_dict: Dict[str, Any]) -> T:
        d = src_dict.copy()
        username = d.pop("username", UNSET)

        course = d.pop("course", UNSET)

        enrollment_json = cls(
            username=username,
            course=course,
        )

        enrollment_json.additional_properties = d
        return enrollment_json

    @property
    def additional_keys(self) -> List[str]:
        return list(self.additional_properties.keys())

    def __getitem__(self, key: str) -> Any:
        return self.additional_properties[key]

    def __setitem__(self, key: str, value: Any) -> None:
        self.additional_properties[key] = value

    def __delitem__(self, key: str) -> None:
        del self.additional_properties[key]

    def __contains__(self, key: str) -> bool:
        return key in self.additional_properties
